rename scenario example test
